colorprofiles: split the profile vocabulary out of the module API
The enums (dt_colorspaces_color_profile_type_t, dt_iop_color_intent_t, the direction
mask, the CICP triple) are serialised into iop params in the library DB and into XMP
sidecars: frozen ABI, and by far the most-wanted thing in this header. Almost every
consumer wants only a profile type to store in its params and an intent to pass along.

Carrying that vocabulary in the same header as the module's API meant <lcms2.h> and
<pthread.h> reached 300 of 425 translation units, nearly all of which call neither.

colorprofiles/profile_types.h now holds the vocabulary and depends on <glib.h> alone.
Measured: translation units transitively seeing <lcms2.h> 300 -> 255. The remaining 255
are the ones that will be converted call site by call site; the header split is what
makes "does this file actually do colour management?" answerable from its includes.

Thirteen includers are re-pointed at whichever header they actually need -- six at the
vocabulary alone, and seven that had been reaching the vocabulary through some other
header without ever naming it. Four of the thirteen are themselves headers
(colorprofiles/iop_profile.h, develop/iop_profile.h, common/exif.h,
common/mipmap_cache.h, common/cups_print.h, imageio/imageio_jpeg.h,
control/jobs/control_jobs.h ...), which is where the transitive fan-out came from.

A caveat about the tool, because it cost a red build here: tools/header_consumers.py's
"files using nothing from it" bucket means "this include is redundant, the file reaches
those symbols another way" -- NOT "the file does not use them". Dropping those seven
outright still compiled in Debug and in Release, and broke build-nofeatures, where
control/jobs/control_jobs.h lost the two types dt_control_export() is declared with.
Under this tree's rule that every file names what it uses, redundant is not the same as
removable: all seven now include the vocabulary directly.

Two details worth naming:

  - dt_iop_color_intent_t borrowed lcms2's INTENT_* constants for its values, which
    would have dragged <lcms2.h> into the vocabulary header on its own. It is spelled
    with literals now; the values are fixed by the ICC specification, and colorspaces.c
    -- the one place that sees both definitions -- static-asserts they still agree.

  - imageio_format_api.h is an X-macro header with no include guard, re-included in the
    same translation unit and expanded inside struct bodies. Its include already sat
    inside the #ifdef FULL_API_H block, so it was swapped in place.

dt_colorspaces_is_matrix_profile_type() and _is_embedded_or_matrix_profile_type() had no
callers anywhere and are deleted rather than moved. Verified by symbol-set diff across
the split: those two are the only declarations that disappeared.

Built clean in all four local configurations: Release, Debug, clang Debug, nofeatures.
