+++ src/colorprofiles/colorspaces.h
@@ -40,6 +40,7 @@
 #ifndef DT_COLORPROFILES_COLORSPACES_H
 #define DT_COLORPROFILES_COLORSPACES_H
 
+#include "colorprofiles/profile_types.h"
 #include "math/matrices.h"
 #include "system/simd.h"
 
@@ -57,125 +58,14 @@ typedef struct _GtkWidget GtkWidget;
 extern "C" {
 #endif
 
+// max samples in a tone-curve LUT built from a profile
+#define LUT_SAMPLES 0x10000
+
 // this was removed from lcms2 in 2.4
 #ifndef TYPE_XYZA_FLT
   #define TYPE_XYZA_FLT (FLOAT_SH(1)|COLORSPACE_SH(PT_XYZ)|EXTRA_SH(1)|CHANNELS_SH(3)|BYTES_SH(4))
 #endif
 
-// max iccprofile file name length
-#define DT_IOP_COLOR_ICC_LEN 512
-#define LUT_SAMPLES 0x10000
-
-
-// constants fit to the ones from lcms.h:
-typedef enum dt_iop_color_intent_t
-{
-  DT_INTENT_PERCEPTUAL = INTENT_PERCEPTUAL,                       // 0
-  DT_INTENT_RELATIVE_COLORIMETRIC = INTENT_RELATIVE_COLORIMETRIC, // 1
-  DT_INTENT_SATURATION = INTENT_SATURATION,                       // 2
-  DT_INTENT_ABSOLUTE_COLORIMETRIC = INTENT_ABSOLUTE_COLORIMETRIC, // 3
-  DT_INTENT_LAST
-} dt_iop_color_intent_t;
-
-typedef enum dt_colorspaces_profile_type_t
-{
-  DT_COLORSPACES_PROFILE_TYPE_INPUT = 1,
-  DT_COLORSPACES_PROFILE_TYPE_WORK = 2,
-  DT_COLORSPACES_PROFILE_TYPE_EXPORT = 3,
-  DT_COLORSPACES_PROFILE_TYPE_DISPLAY = 4,
-  DT_COLORSPACES_PROFILE_TYPE_SOFTPROOF = 5
-} dt_colorspaces_profile_type_t;
-
-typedef enum dt_colorspaces_color_profile_type_t
-{
-  DT_COLORSPACE_NONE = -1,
-  DT_COLORSPACE_FILE = 0,
-  DT_COLORSPACE_SRGB = 1,
-  DT_COLORSPACE_ADOBERGB = 2,
-  DT_COLORSPACE_LIN_REC709 = 3,
-  DT_COLORSPACE_LIN_REC2020 = 4,
-  DT_COLORSPACE_XYZ = 5,
-  DT_COLORSPACE_LAB = 6,
-  DT_COLORSPACE_INFRARED = 7,
-  DT_COLORSPACE_DISPLAY = 8,
-  DT_COLORSPACE_EMBEDDED_ICC = 9,
-  DT_COLORSPACE_EMBEDDED_MATRIX = 10,
-  DT_COLORSPACE_STANDARD_MATRIX = 11,
-  DT_COLORSPACE_ENHANCED_MATRIX = 12,
-  DT_COLORSPACE_VENDOR_MATRIX = 13,
-  DT_COLORSPACE_ALTERNATE_MATRIX = 14,
-  DT_COLORSPACE_BRG = 15,
-  DT_COLORSPACE_EXPORT = 16, // export and softproof are categories and will return NULL with dt_colorspaces_get_profile()
-  DT_COLORSPACE_SOFTPROOF = 17,
-  DT_COLORSPACE_WORK = 18,
-  DT_COLORSPACE_DISPLAY2 = 19,
-  DT_COLORSPACE_REC709 = 20,
-  DT_COLORSPACE_PROPHOTO_RGB = 21,
-  DT_COLORSPACE_PQ_REC2020 = 22,
-  DT_COLORSPACE_HLG_REC2020 = 23,
-  DT_COLORSPACE_PQ_P3 = 24,
-  DT_COLORSPACE_HLG_P3 = 25,
-  DT_COLORSPACE_ITUR_BT1886 = 26,
-  DT_COLORSPACE_DISPLAY_P3 = 27,
-  DT_COLORSPACE_LAST = 28
-} dt_colorspaces_color_profile_type_t;
-
-typedef enum dt_colorspaces_color_mode_t
-{
-  DT_PROFILE_NORMAL = 0,
-  DT_PROFILE_SOFTPROOF,
-  DT_PROFILE_GAMUTCHECK
-} dt_colorspaces_color_mode_t;
-
-typedef enum dt_colorspaces_profile_direction_t
-{
-  DT_PROFILE_DIRECTION_IN = 1 << 0,
-  DT_PROFILE_DIRECTION_OUT = 1 << 1,
-  DT_PROFILE_DIRECTION_DISPLAY = 1 << 2,
-  DT_PROFILE_DIRECTION_CATEGORY = 1 << 3, // categories will return NULL with dt_colorspaces_get_profile()
-  DT_PROFILE_DIRECTION_WORK = 1 << 4,
-  DT_PROFILE_DIRECTION_DISPLAY2 = 1 << 5,
-  DT_PROFILE_DIRECTION_ANY = DT_PROFILE_DIRECTION_IN | DT_PROFILE_DIRECTION_OUT | DT_PROFILE_DIRECTION_DISPLAY
-                             | DT_PROFILE_DIRECTION_CATEGORY
-                             | DT_PROFILE_DIRECTION_WORK
-                             | DT_PROFILE_DIRECTION_DISPLAY2
-} dt_colorspaces_profile_direction_t;
-
-/* CICP color primaries (Recommendation ITU-T H.273) */
-typedef enum dt_colorspaces_cicp_color_primaries_t
-{
-    DT_CICP_COLOR_PRIMARIES_REC709 = 1,
-    DT_CICP_COLOR_PRIMARIES_UNSPECIFIED = 2,
-    DT_CICP_COLOR_PRIMARIES_REC2020 = 9,
-    DT_CICP_COLOR_PRIMARIES_XYZ = 10,
-    DT_CICP_COLOR_PRIMARIES_P3 = 12 // D65
-} dt_colorspaces_cicp_color_primaries_t;
-
-/* CICP transfer characteristics (Recommendation ITU-T H.273) */
-typedef enum dt_colorspaces_cicp_transfer_characteristics_t
-{
-    DT_CICP_TRANSFER_CHARACTERISTICS_REC709 = 1,
-    DT_CICP_TRANSFER_CHARACTERISTICS_UNSPECIFIED = 2,
-    DT_CICP_TRANSFER_CHARACTERISTICS_REC601 = 6,
-    DT_CICP_TRANSFER_CHARACTERISTICS_LINEAR = 8,
-    DT_CICP_TRANSFER_CHARACTERISTICS_SRGB = 13,
-    DT_CICP_TRANSFER_CHARACTERISTICS_REC2020_10B = 14,
-    DT_CICP_TRANSFER_CHARACTERISTICS_REC2020_12B = 15,
-    DT_CICP_TRANSFER_CHARACTERISTICS_PQ = 16,
-    DT_CICP_TRANSFER_CHARACTERISTICS_HLG = 18
-} dt_colorspaces_cicp_transfer_characteristics_t;
-
-/* CICP matrix coefficients (Recommendation ITU-T H.273) */
-typedef enum dt_colorspaces_cicp_matrix_coefficients_t
-{
-    DT_CICP_MATRIX_COEFFICIENTS_IDENTITY = 0,
-    DT_CICP_MATRIX_COEFFICIENTS_REC709 = 1,
-    DT_CICP_MATRIX_COEFFICIENTS_UNSPECIFIED = 2,
-    DT_CICP_MATRIX_COEFFICIENTS_SYCC = 5,
-    DT_CICP_MATRIX_COEFFICIENTS_REC601 = 6,
-    DT_CICP_MATRIX_COEFFICIENTS_REC2020_NCL = 9,
-    DT_CICP_MATRIX_COEFFICIENTS_CHROMA_DERIVED_NCL = 12
-} dt_colorspaces_cicp_matrix_coefficients_t;
 
 typedef struct dt_colorspaces_t
 {
@@ -386,32 +276,7 @@ void dt_colorspaces_cygm_to_rgb(float *out, int num, double CAM_to_RGB[3][4]);
 void dt_colorspaces_rgb_to_cygm(float *out, int num, double RGB_to_CAM[4][3]);
 
 
-static inline dt_colorspaces_color_profile_type_t sanitize_colorspaces(dt_colorspaces_color_profile_type_t colorspace)
-{
-  // Remap unused colorspaces to valid ones
-  if(colorspace == DT_COLORSPACE_DISPLAY2)
-    return DT_COLORSPACE_DISPLAY;
-  else
-    return (dt_colorspaces_color_profile_type_t)MIN(colorspace, DT_COLORSPACE_LAST - 1);
-}
 
-static inline gboolean dt_colorspaces_is_raw_matrix_profile_type(const dt_colorspaces_color_profile_type_t type)
-{
-  return (type == DT_COLORSPACE_STANDARD_MATRIX
-          || type == DT_COLORSPACE_ENHANCED_MATRIX
-          || type == DT_COLORSPACE_VENDOR_MATRIX
-          || type == DT_COLORSPACE_ALTERNATE_MATRIX);
-}
-
-static inline gboolean dt_colorspaces_is_matrix_profile_type(const dt_colorspaces_color_profile_type_t type)
-{
-  return dt_colorspaces_is_raw_matrix_profile_type(type) || type == DT_COLORSPACE_EMBEDDED_MATRIX;
-}
-
-static inline gboolean dt_colorspaces_is_embedded_or_matrix_profile_type(const dt_colorspaces_color_profile_type_t type)
-{
-  return (type == DT_COLORSPACE_EMBEDDED_ICC) || dt_colorspaces_is_matrix_profile_type(type);
-}
