+++ src/colorprofiles/iop_profile.h
@@ -42,7 +42,7 @@
 
 #include "pixel/format.h"   // dt_iop_colorspace_type_t
 #include "common/colorspaces_inline_conversions.h"
-#include "colorprofiles/colorspaces.h"
+#include "colorprofiles/profile_types.h"
 /* develop/imageop.h is deliberately NOT included: every iop/pixelpipe type used
  * below is used as a POINTER and already tag-declared a few lines down. Including
  * it made common/ depend on develop/ (a layering inversion) and closed a 6-node
