+++ src/common/cups_print.h
@@ -24,7 +24,7 @@
 #ifndef DT_COMMON_CUPS_PRINT_H
 #define DT_COMMON_CUPS_PRINT_H
 
-#include "colorprofiles/colorspaces.h"
+#include "colorprofiles/profile_types.h"
 
 #define MAX_NAME 128
