+++ src/develop/pixelpipe_hb.c
@@ -51,7 +51,7 @@
     You should have received a copy of the GNU General Public License
     along with darktable.  If not, see <http://www.gnu.org/licenses/>.
 */
-#include "colorprofiles/colorspaces.h"
+#include "colorprofiles/profile_types.h"
 #include "system/capabilities.h"
 #include "system/sys_resources.h"
 #include "common/global_mutexes.h"
