+++ src/imageio/imageio_tiff.c
@@ -33,8 +33,8 @@
     You should have received a copy of the GNU General Public License
     along with darktable.  If not, see <http://www.gnu.org/licenses/>.
 */
-#include "imageio_tiff.h"
 #include "colorprofiles/colorspaces.h"
+#include "imageio_tiff.h"
 #include "system/macros.h"
 #include "system/mem_alloc.h"
 #include "common/logging.h"
